fix autoimport in updaters
